requirements.txt
anthropic>=0.92.0
google-genai>=1.0.0
python-dotenv>=1.0.0
